Have sphinx copy the markdown files for use in RTD
Trying to find a way to include my markdown docs in the root directory
in the readthedocs TOC. Unfortunately, apparently RST is unable to find
them in containing directories (`..`) unless I use a `.. include:: `
directive, which only seems to work with rst and not markdown. Copying
the files to a subdir in the Makefile works but means they may not
contain the latest changes to documentation (should I forget to `make
docs`). Symlinks don't seem to be working.

This commit makes a `setup` function in `conf.py` that tries to copy the
markdown files into a subdir, `md`, which I have `.gitignore`d (since I
don't need the duplicate files). Hopefully readthedocs will use this to
copy the files to a subdir on each build and therefore have the most
current markdown docs available.