Add ability to manually specify IP address
Fix for https://github.com/n8henrie/fauxmo/issues/1
Now users can manually specify an IP address in config.json if they are
encountering an issue with the automatic detection. If the address is
set to "auto" (or if this part of the config is absent) it will continue
to function as before, hopefully not affecting existing users.

Also added some extra debugging calls related to ip_address stuff and
removed the unused `DEBUG` key from the config.