Add additional HTTP verbs / options in RestApiHandler
Expanded the potential of RestApiHandler by modifying it to accept other
HTTP verbs, POST data, and allowing it to have different data sent for
`on()` and `off()` methods. This will mean most users now have a bit of
redundancy in their configs but should greatly expand the number of
devices that `RestAPIHandler` can interact with.

Breaking changes
===

- RestApiHandler's `json` kwarg is now either `on_json` or `off_json`,
  to allow different JSON bodies to be sent for `on()` or `off()`

Additions
---

- Now also accepts `on_data`, `off_data` for POST bodies
- Will now work with other HTTP verbs like `put`
- Now supports HTTP Basic and Digest authentication
- Reduced redundancy in `on()` and `off()` methods with
  `functools.partialmethod`