Update tests for asyncio
For testing, use `multiprocessing` to create a Fauxmo server in a
thread with a pytest `yield_fixture`, which we can then use to test our
UDP and TCP servers using `socket` and `requests`, respectively. Of
note, the `yield_fixture` is only creating the server once for the
entire testing session (`scope=("session")`), so while a brief
(currently 1 second) delay is required to allow Fauxmo time to set up
the server, this is only required once for all tests. Also had to
include a 1 second timeout for the UDP server to respond for some reason
(in `test_udp_search`).