Don't decode the incoming UDP broadcasts
Closes #7. Fauxmo was trying to decode bytes from incoming UDP
broadcasts that were likely being sent by devices other than the Echo
and perhaps not in the system's default encoding, leading to a
UnicodeDecodeError. With this commit, Fauxmo no longer tries to decode
the incoming UDP message, since it doesn't need to be decoded to search
for the critical components that help identify the Echo's device search.
Also updated 1 test to match the new behavior.