Make logging.INFO more informative
Change log level for several debug messages so that logging.INFO (`-vv`)
shows basic info about what is turning on and when, as well as what the
apparent device status is. Also use the device name in the info output
(but still include the plugin type in the debug output, in case there is
a problem with the name).