Adds a close function to plugins in a backwards-compatible way.
The main loop in fauxmo.py now keeps track of the plugin instances as
it creates them, and then calls a new close() method on each instance
when tearing down. The close method was added to FauxmoPlugin as a
non-abstract method, so existing plugins should not need to change.

Also added a minimal test. All this proves at the moment is that it
is possible to call close() on a plugin instance without negative
effects.