Add support for reporting the latest action as state
Adds an implementation of `get_state` into the FauxmoPlugin abstract
base class that can be opted into by returning `super().get_state()` in
a plugin's `get_state()` method. This behavior should be configured with
a `use_fake_state` kwarg, alerting users that the state reported will
not necessarily be the actual device state.

Also do not report success for `get_state` if the returned value is not
either "on" or "off" (the only values that will be understood by the
Echo as far as I know).

Fixes n8henrie/fauxmo#80