Add `<serialNumber>` to Fauxmo's response
[pywemo](https://github.com/pavoni/pywemo/) seems to use this bit of xml
to set a [device attribute called `.serialNumber`][0].

HomeAssistant in turn deduplicates discovered devices based on this
attribute][1].

Without it set in XML, HomeAssistant only discovers the *first* Fauxmo
device, assuming the remainder are duplicates (since they have the same
serial number).

This commit adds the serialNumber to the XML, and also adds the location
(which includes host and port) to the generation of the serial number,
just in case there were two devices with the same name for some reason.
Probably unnecessary, but shouldn't cause harm as far as I can tell.

Fixes https://github.com/n8henrie/fauxmo/issues/88

[0]: https://github.com/pavoni/pywemo/blob/c32a6d0dbf7c0aee4935f778fd683aa59db3f073/pywemo/ouimeaux_device/api/xsd/device.py#L730
[1]: https://github.com/home-assistant/home-assistant/blob/7b86f0f9265dc8cdcf7ed92c0db7340604134569/homeassistant/components/wemo/__init__.py#L166