Update simplehttpplugin.py
Added a 2s sleep in get_state.  Without this, when Alex turns on the device and them immediately queries the state, and error occurs probably because either the RPI or the Alexa is not ready for an immediate response.  How awful, without the sleep (or just 1s) the response setup fails.  Surely there is a more elegant solution, but this gets the job done at least until somebody has time to work out a better solution.