Add support for timeouts and `shell=True` commands
This commit adds timeouts for commandline, homeassistant, and simplehttp
plugins. Previously blocking code (like a delayed web request or a hung
command) would hang the entire fauxmo process. Ultimately it might be
better to refactor the plugins to all be `async`, but for the moment
adding timeouts seems like a fair workaround.

This commit also adds support for a `shell` keyword for
`CommandLinePlugin`, which can run a less safe command like `sleep 2;
echo foo >> bar.txt'.