Add `initial_state` kwarg for plugins
As of Aug 2023, Alexa has a somewhat new behavior in which it first
polls for device state *before* adding a new device.

This is problematic in a few cases:
- devices that can't report state, which may be relying on having a
  successful `.on()` or `.off()` to set the `_latest_action` and take
  advantage of `use_fake_state`
- devices whose state may not be reportable on demand (perhaps they are
  polling for state changes every 15 minutes or so) and therefore state
  is not yet available

For such cases I've added an optional `initial_state` kwarg, which
should allow `use_fake_state` to have something to report.

See some discussions at:

- https://github.com/n8henrie/fauxmo-plugins/issues/26
- https://github.com/n8henrie/fauxmo/issues/117